renamed from RPi specific . Removed rounding
