+++ src/devices/CpuTemperature/README.md
@@ -1,0 +1,11 @@
+# RpiCpuTemperature
+
+Device bindings for the CPU Temperature Sensor
+
+## Summary
+
+Returns the current temperature of the CPU Temperature Sensor. Useful telemetry in its own right, but also useful for calibrating the Raspberry Pi Sense HAT.
+
+## Binding Notes
+
+## References
