+++ src/devices/CpuTemperature/samples/CpuTemperature.Samples.csproj
@@ -6,7 +6,7 @@
   </PropertyGroup>
 
   <ItemGroup>
-    <ProjectReference Include="../RpiCpuTemperature.csproj" />
+    <ProjectReference Include="../CpuTemperature.csproj" />
   </ItemGroup>
 
 </Project>
