--- src/devices/RpiCpuTemperature/README.md
@@ -1,11 +1,0 @@
-# RpiCpuTemperature
-
-Device bindings for the Raspberry Pi CPU Temperature Sensor
-
-## Summary
-
-Returns the current temperature of the Raspberry Pi Temperature Sensor. Useful telemetry in its own right, but also useful for calibrating the Raspberry Pi Sense HAT.
-
-## Binding Notes
-
-## References
